add necessary imports in XinglongHRS pipeline
