+++ gamse/pipelines/xinglong216hrs/reduce_doublefiber.py
@@ -1,4 +1,20 @@
 import os
+import re
+import logging
+logger = logging.getLogger(__name__)
+
+import numpy as np
+import astropy.io.fits as fits
+
+from ...echelle.imageproc import combine_images, array_to_table, fix_pixels
+from ...echelle.trace import find_apertures, load_aperture_set, TraceFigureCommon
+from ...echelle.flat  import get_fiber_flat, mosaic_flat_auto, mosaic_images
+from ...echelle.extract import extract_aperset
+from ...echelle.wlcalib import (wlcalib, recalib, get_calib_from_header,
+                               get_time_weight, find_caliblamp_offset,
+                               reference_wavelength,
+                               reference_self_wavelength)
+from ..common import plot_background_aspect1, FormattedInfo
 
 def reduce_doublefiber(logtable, config):
     """Reduce the multi-fiber data of Xinglong 2.16m HRS.
